+++ src/model/IPriorityQueue.java
@@ -2,7 +2,7 @@ package model;
 
 public interface IPriorityQueue<T> {
 
-    void insert(T obj, int key) throws Exception;
+    void insert(T obj, double key) throws Exception;
 
     public T extract() throws Exception;
